backend/pyproject.toml
[project]
name = "mirofish-offline-backend"
version = "0.2.0"
description = "MiroFish-Offline - Offline-first fork running on Neo4j + Ollama"
requires-python = ">=3.11"
license = { text = "AGPL-3.0" }
authors = [
    { name = "MiroFish Team" }
]

dependencies = [
    # 核心框架
    "flask>=3.0.0",
    "flask-cors>=6.0.0",
    
    # LLM 관련
    "openai>=1.0.0",
    "claude-agent-sdk>=0.1.0",
    "anyio>=4.0.0",
    
    # Neo4j graph database driver
    "neo4j>=5.15.0",
    
    # OASIS 社交媒体模拟
    "camel-oasis==0.2.5",
    "camel-ai==0.2.78",
    
    # 文件处理
    "PyMuPDF>=1.24.0",
    # 编码检测（支持非UTF-8编码的文本文件）
    "charset-normalizer>=3.0.0",
    "chardet>=5.0.0",
    
    # 工具库
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pipreqs>=0.5.0",
]

[tool.uv]
override-dependencies = [
    "pillow>=11.0.0",
    "tiktoken>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
